volume/system-under-test/requirements.txt
fake-bpy-module-latest==20250522
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
